Actualizacion de logica para archivos AFP: Renombramiento de archivos
